Add authentication and authorization to the application

+++ src/main/java/com/grocery/repository/WishlistRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.grocery.model.Wishlist;
 
-public interface WishlistRepository extends CrudRepository<Wishlist, Integer> {
+public interface WishlistRepository extends CrudRepository<Wishlist, Long> {
 
 }
